linebot_mo2/requirements.txt
line-bot-sdk
flask
gunicorn
